Feature/#30 reply

+++ community/src/main/java/com/nocapstone/community/dto/CommunityData.kt
@@ -5,12 +5,12 @@ data class CommunityData(
 )
 
 data class Post(
-    val postId: Int,
+    val postId: Long,
     val title: String,
     val thumbnail: String,
     val replyCount : Int
 )
 
 data class PostId(
-    val postId: Int
+    val postId: Long
 )
